refactory config rw
